chore(engine): Refactor action method names to be the same as export.
