+++ packages/engine/src/actions/createGetGlobal.js
